Create function that acts as computer opponent
The goal of this function is for it to randomly
return one of three sting values: "rock", "paper",
or "scissors". A random number is generated between
0 and 0.99. If the number is less than 0.33, the
function returns "rock". If greater than or equal
to 0.33 and less than 0.66, it returns "paper". If
greater than or equal to 0.67, it returns
"scissors".