Create function that gets player response
Function getHumanChoice() will prompt the user
to enter either rock, paper, or scissors. It turns
their response into a string of all uppercase
characters so that responses like "Rock", "rock",
and "rOcK" will all be turned into "ROCK". If the
player response is either "ROCK", "PAPER", or
"SCISSORS", the function will log the player's
choice back to them on the console and return
either integer 1 for rock, integer 2 for paper,
and integer 3 for scissors. If the player responds
with something other than rock, paper, or scissors,
it will continue to prompt the user till a valid
response is received.