Change all function return variables
Return values for getComputerChoice() were changed
from "rock", "paper", and "scissors" to "ROCK",
"PAPER", and "SCISSORS". Return values for
getHumanChoice() function were changed from 1, 2,
and 3 to "ROCK", "PAPER", and "SCISSORS" to make
it more intuitive while reading the code to
determine a winner when comparing the return values
of getComputerChoice() and getHumanChoice().